Anticipating problems with case sensitivity
In the default layout, change the import of the Header component to use consistent casing with the actual file name. This is done to maintain consistency and prevent potential issues with case-sensitive filesystems, as BTRFS and EXT4.
Windows would not be affected.
See https://www.collabora.com/news-and-blog/blog/2020/08/27/using-the-linux-kernel-case-insensitive-feature-in-ext4/